@@ -20,24 +20,26 @@
   },
   "homepage": "https://github.com/modernmaster/forecaster",
   "dependencies": {
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4",
     "react-router-dom": "^4.3.1",
-    "react-web-notification": "^0.3.2",
-    "rest": "^1.3.1",
-    "sockjs-client": "^1.0.3",
+    "react-web-notification": "^0.5.0",
+    "rest": "^2.0.0",
+    "sockjs-client": "^1.3.0",
     "stompjs": "^2.3.3",
     "svg-inline-react": "^3.1.0",
-    "webpack": "^1.12.2"
+    "webpack": "^4.29.6"
   },
   "scripts": {
     "watch": "webpack --watch -d"
   },
   "devDependencies": {
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.2.7",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0"
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.3.4",
+    "@babel/preset-env": "^7.3.4",
+    "@babel/preset-react": "^7.0.0",
+    "babel-loader": "^8.0.5",
+    "babel-polyfill": "^6.26.0",
+    "webpack-cli": "^3.2.3"
   }
 }
